+++ R/manip.r
@@ -2,7 +2,7 @@
 #'
 #' @family single table verbs
 #' @param .data A tbl. All main verbs are S3 generics and provide methods
-#'   for \code{\link{tbl_df}}, \code{\link{tbl_dt}} and \code{\link{tbl_sql}}.
+#'   for \code{\link{tbl_df}}, \code{\link[dtplyr]{tbl_dt}} and \code{\link{tbl_sql}}.
 #' @param ... Logical predicates. Multiple conditions are combined with \code{&}.
 #' @param .dots Used to work around non-standard evaluation. See
 #'   \code{vignette("nse")} for details.
@@ -39,7 +39,7 @@ filter_ <- function(.data, ..., .dots) {
 #'
 #' @family single table verbs
 #' @param .data A tbl. All main verbs are S3 generics and provide methods
-#'   for \code{\link{tbl_df}}, \code{\link{tbl_dt}} and \code{\link{tbl_sql}}.
+#'   for \code{\link{tbl_df}}, \code{\link[dtplyr]{tbl_dt}} and \code{\link{tbl_sql}}.
 #' @param ... Integer row values
 #' @inheritParams filter
 #' @export
@@ -95,7 +95,7 @@ slice_ <- function(.data, ..., .dots) {
 #'
 #' \dontrun{
 #' # You can't with data tables or databases
-#' by_cyl_dt <- mtcars %>% tbl_dt() %>% group_by(cyl)
+#' by_cyl_dt <- mtcars %>% dtplyr::tbl_dt() %>% group_by(cyl)
 #' by_cyl_dt %>% summarise(a = n(), b = a + 1)
 #'
 #' by_cyl_db <- src_sqlite(":memory:", create = TRUE) %>%
@@ -174,7 +174,8 @@ transmute_.default <- function(.data, ..., .dots) {
 
 #' Arrange rows by variables.
 #'
-#' Use \code{\link{desc}} to sort a variable in descending order.
+#' Use \code{\link{desc}} to sort a variable in descending order. Generally,
+#' this will not also automatically order by grouping variables.
 #'
 #' @section Locales:
 #'
@@ -213,22 +214,7 @@ arrange_ <- function(.data, ..., .dots) {
 #' As well as using existing functions like \code{:} and \code{c}, there are
 #' a number of special functions that only work inside \code{select}
 #'
-#' \itemize{
-#'  \item \code{starts_with(x, ignore.case = TRUE)}:
-#'    names starts with \code{x}
-#'  \item \code{ends_with(x, ignore.case = TRUE)}:
-#'    names ends in \code{x}
-#'  \item \code{contains(x, ignore.case = TRUE)}:
-#'    selects all variables whose name contains \code{x}
-#'  \item \code{matches(x, ignore.case = TRUE)}:
-#'    selects all variables whose name matches the regular expression \code{x}
-#'  \item \code{num_range("x", 1:5, width = 2)}:
-#'    selects all variables (numerically) from x01 to x05.
-#'  \item \code{one_of("x", "y", "z")}:
-#'    selects variables provided in a character vector.
-#'  \item \code{everything()}:
-#'    selects all variables.
-#' }
+
 #'
 #' To drop variables, use \code{-}. You can rename variables with
 #' named arguments.